extract to public outer classes
