+++ jvm/ml4ir-inference/src/main/java/ml4ir/inference/tensorflow/utils/InputFeatures.java
@@ -1,0 +1,3 @@
+package ml4ir.inference.tensorflow.utils;
+
+public class InputFeatures extends BaseFeatures {}
