+++ jvm/ml4ir-inference/src/main/java/ml4ir/inference/tensorflow/utils/ModelFeatures.java
@@ -47,82 +47,3 @@ public class ModelFeatures {
     }
 }
 
-class BaseFeatures {
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("node_name")
-    private String nodeName;
-
-    @JsonProperty("trainable")
-    private boolean isTrainable;
-
-    @JsonProperty("dtype")
-    private String dtype;
-
-    @JsonProperty("log_at_inference")
-    private boolean logAtInference;
-
-    @JsonProperty("feature_layer_info")
-    private FeatureLayerInfo featureLayerInfo;
-
-    @JsonProperty("serving_info")
-    private ServingInfo servingInfo;
-
-    @JsonProperty("tfrecord_type")
-    private String tfRecordType;
-
-    @JsonProperty("default_value")
-    private String defaultValue;
-
-    public String getNodeName() {
-        return nodeName;
-    }
-
-    public String getDtype() {
-        return dtype;
-    }
-
-    public ServingInfo getServingInfo() {
-        return servingInfo;
-    }
-
-    public String getTfRecordType() {
-        return tfRecordType;
-    }
-
-    public String getDefaultValue() {
-        return defaultValue;
-    }
-}
-
-class FeatureLayerInfo {
-    @JsonProperty("type")
-    private String type;
-
-    @JsonProperty("shape")
-    private String shape;
-
-    @JsonProperty("max_length")
-    private int maxLength;
-}
-
-class ServingInfo {
-    @JsonProperty("name")
-    private String name;
-
-    @JsonProperty("required")
-    private boolean isRequired;
-
-    public String getName() {
-        return name;
-    }
-}
-
-class QueryKey extends BaseFeatures {}
-
-class Rank extends BaseFeatures {}
-
-class Label extends BaseFeatures {}
-
-class InputFeatures extends BaseFeatures {}
