+++ jvm/ml4ir-inference/src/main/java/ml4ir/inference/tensorflow/utils/ServingInfo.java
@@ -1,0 +1,15 @@
+package ml4ir.inference.tensorflow.utils;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public class ServingInfo {
+    @JsonProperty("name")
+    private String name;
+
+    @JsonProperty("required")
+    private boolean isRequired;
+
+    public String getName() {
+        return name;
+    }
+}
